Workaround for the OverlayFS bug present before Kernel 4.2
When a process has a file open in an OverlayFS directory,
the information in /proc/<pid>/fd/<fd> and /proc/<pid>/fdinfo/<fd>
is wrong, so, if --overlayfs is specified, we grab that information
from the mountinfo table instead.

This is done every time fill_fdlink is called. We first enter the
mount namespace of the process and check to see if there are any overlayFS
mounted directories in the mountinfo table. If so, we concatenate the
mountpoint with the name of the file, and stat the resulting path to
check if we found the correct device id and node number. If that is
the case, we update the mount id and link variables with the
correct values.

This update is skipped if there is an entry in the
mountinfo table with the same known device number and
mount id, signifying that we already have the correct id.

At the end of this procedure, we exit the mount namespace
of the process, back to the original one.